refactor: separate XML output destinations
